requirements.txt
# Core ML
torch==2.2.2
torchvision==0.17.2

# Image handling
Pillow==10.3.0

# Data handling
numpy==1.26.4

# Optional but recommended for training / evaluation
scikit-learn==1.5.1
matplotlib==3.9.0
tqdm==4.66.4

# Optional for notebooks
ipython>=8.10,<9